Send event notification for overlapping time range
Previously, we were sending push notifications of new events added only
to users whose time range preferences fully contains the event start
and end times. Now we send to users whose time range preferences
overlaps at all with the event time. The justification is that people
can still be interested and come and go as they please, and do not need
to be present the entire duration of the event.